feat: extend liquid glass effect to all pages
- Add reusable GlassOrbs component for decorative background orbs
- Apply glass orbs to: home, about, contact, iris, TOS, privacy
- Upgrade contact card to liquid glass styling
- Add glass-styled sections to about page
- Upgrade iris auth card to liquid glass
- LiquidGlassHero now wraps shared GlassOrbs component
